tests/test_functions_database.py
import streamlit as st
import sys
import os

sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
from src.core.database import get_categorias_erro

print(get_categorias_erro())
