requirements-gatekeep.txt
requests
tomli
supabase
python-dotenv
google-generativeai